Refactor directory structure and delete unused files
