+++ .gitignore
@@ -1,3 +1,8 @@
+# Visual Studio Code settings
+.vscode/
+
+.DS_STORE
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
