+++ backend/agent/state.py
@@ -1,7 +1,7 @@
 from typing import TypedDict, Annotated, Dict, Any, List
 
 # from langgraph.graph import add_messages
-from backend.src.agent.utils import add_messages_to_dict
+from backend.agent.utils import add_messages_to_dict
 from langchain_core.messages import AnyMessage
