--- frontend/modal/graph.txt
@@ -1,15 +1,0 @@
-                                                                +-----------+                                                        
-                                                                | __start__ |                                                        
-                                                                +-----------+                                                        
-                                                                      .                                                              
-                                                                      .                                                              
-                                                                      .                                                              
-                                                              +---------------+                                                      
-                                                          ....| General Agent |.....                                                 
-                                                ..........    +---------------+.    ...........                                      
-                                      ..........            ...                 .....          ...........                           
-                            ..........                   ...                         ...                  ...........                
-                       .....                           ..                               ...                          ......          
-+-------------------------+           +--------------------------+           +-------------------------------+           +---------+ 
-| General Agent_rag_tools |           | General Agent_safe_tools |           | General Agent_sensitive_tools |           | __end__ | 
-+-------------------------+           +--------------------------+           +-------------------------------+           +---------+
