+++ graph.txt
@@ -6,10 +6,10 @@
                                                                       .                                                              
                                                               +---------------+                                                      
                                                           ****| General Agent |.....                                                 
-                                                **********    +---------------+*    ...........                                      
-                                      **********            ...                 *****          ...........                           
-                            **********                   ...                         ***                  ...........                
-                       *****                           ..                               ***                          ......          
+                                                **********    +---------------+.    ...........                                      
+                                      **********            ***                 .....          ...........                           
+                            **********                   ***                         ...                  ...........                
+                       *****                           **                               ...                          ......          
 +-------------------------+           +--------------------------+           +-------------------------------+           +---------+ 
 | General Agent_rag_tools |           | General Agent_safe_tools |           | General Agent_sensitive_tools |           | __end__ | 
 +-------------------------+           +--------------------------+           +-------------------------------+           +---------+
